Cuarto Ingenieria Informatica (Plan 40)/Herr. Prof. de Desarrollo de Software/Proyecto/HPDS_Coworking_Manager/CoworkingManager/src/main/java/es/ulpgc/hpds/conditions/IsAvaliableCondition.java
package es.ulpgc.hpds.conditions;

import es.ulpgc.hpds.environment.Sensor;
import es.ulpgc.hpds.persistence.Condition;

import java.util.ArrayList;

public class IsAvaliableCondition implements Condition {
    ArrayList<Sensor> sensorsAssociated;

    public IsAvaliableCondition(ArrayList<Sensor> sensorsAssociated){
        this.sensorsAssociated=sensorsAssociated;
    }
    @Override
    public boolean evaluate(){
        for (Sensor sensor: sensorsAssociated){
            if(sensor.isTriggered()){
                return false;
            }
        }
        return true;
    }
}

